-add model to benchmark

+++ benchmark/rocm_benchmark.py
@@ -159,7 +159,7 @@ class ROCmBenchmark(Benchmark):
         # encoder-decoder has vocab size saved differently
         vocab_size = config.vocab_size if hasattr(config, "vocab_size") else config.encoder.vocab_size
         print('!!!!! start inference')
-        input_ids = random.randint(key, (batch_size, sequence_length), dtype=jax.numpy.uint64, minval=0,
+        input_ids = random.randint(key, (batch_size, sequence_length), minval=0,
                                    maxval=vocab_size)
         print('!!!!! start inference')
         if self.args.fp16:
